change uint to uint256

+++ contracts/ERC721/Storage.sol
@@ -3,23 +3,23 @@ pragma solidity ^0.4.18;
 contract Storage {
 
     // Total amount of tokens
-    uint public totalTokens;
+    uint256 public totalTokens;
     
     struct Token {
         address tokenOwner;
-        uint tokenPrice;
+        uint256 tokenPrice;
         bytes32 tokenName;
         bytes32 tokenType;
         address tokenApproval;
     }
     //mapping from tokenID to token details
-    mapping (uint => Token) tokens;
+    mapping (uint256 => Token) tokens;
 
     // Mapping from owner to list of owned token IDs
-    mapping (address => uint[]) public ownedTokens;
+    mapping (address => uint256[]) public ownedTokens;
 
     // Mapping from token ID to index of the owner tokens list
-    mapping(uint => uint) public ownedTokensIndex;
+    mapping(uint256 => uint256) public ownedTokensIndex;
 
     // Mapping to allow access to change Storage
     mapping(address => bool) public accessAllowed;
